bugfix: load pacman db from previous release
If we create a new empty .db.tar.gz and later load the
one from previous release, it is older.
pacman only loads newer versions, so it keeps the newer one that is the empty created.

to fix it, we set the date of the empty .db.tar.gz to 1970-01-01.